refactor to type definition method
