+++ vyper/semantics/types/bases.py
@@ -291,6 +291,15 @@ class BaseTypeDefinition:
         """
         return self.abi_type.json_abi_name()
 
+    def json_abi(self, name="") -> Dict:
+        """
+        The JSON ABI output
+        """
+        return {
+            "name": name,
+            "type": self.json_abi_type,
+        }
+
     def from_annotation(self, node: vy_ast.VyperNode, *args: Any, **kwargs: Any) -> None:
         # always raises, user should have used a primitive
         raise StructureException("Value is not a type", node)
