+++ vyper/semantics/types/indexable/sequence.py
@@ -16,6 +16,27 @@ from vyper.semantics.types.value.array_value import BytesArrayDefinition, String
 from vyper.semantics.types.value.numeric import Uint256Definition  # type: ignore
 
 
+def _generate_components_if_struct(type_definition):
+    """
+    Helper function to generate the `components` for a struct in a multidimensional
+    array for JSON ABI.
+
+    Returns the `components` value from `generate_abi_type` of a struct if the
+    array is of struct type. Otherwise, return None.
+    """
+    from vyper.semantics.types.user.struct import StructDefinition
+
+    if isinstance(type_definition, StructDefinition):
+        return type_definition.json_abi()["components"]
+
+    # Early termination if not a struct, static array or dynamic array
+    if not isinstance(type_definition, (ArrayDefinition, DynamicArrayDefinition)):
+        return None
+
+    # Recursively get the value type since it is a static/dynamic array
+    return _generate_components_if_struct(type_definition.value_type)
+
+
 class _SequenceDefinition(IndexableTypeDefinition):
     """
     Private base class for sequence types.
@@ -97,6 +118,33 @@ class ArrayDefinition(_SequenceDefinition):
     def abi_type(self) -> ABIType:
         return ABI_StaticArray(self.value_type.abi_type, self.length)
 
+    def json_abi(self, name=""):
+        from vyper.semantics.types.user.struct import StructDefinition
+
+        if isinstance(self.value_type, StructDefinition):
+            # For structs, set `components` as the struct's components directly
+            return {
+                "name": name,
+                "type": self.json_abi_type,
+                "components": self.value_type.json_abi()["components"],
+            }
+
+        elif isinstance(self.value_type, (ArrayDefinition, DynamicArrayDefinition)):
+            # For static and dynamic arrays, set `components` if the value type is a struct.
+            # Otherwise, the `canonical_abi_type` is sufficient.
+            ret = {
+                "name": name,
+                "type": self.json_abi_type,
+            }
+
+            struct_components = _generate_components_if_struct(self.value_type)
+            if struct_components is not None:
+                ret["components"] = struct_components
+
+            return ret
+
+        return {"name": name, "type": self.json_abi_type}
+
     @property
     def is_dynamic_size(self):
         return self.value_type.is_dynamic_size
@@ -159,6 +207,33 @@ class DynamicArrayDefinition(_SequenceDefinition, MemberTypeDefinition):
     @property
     def abi_type(self) -> ABIType:
         return ABI_DynamicArray(self.value_type.abi_type, self.length)
+
+    def json_abi(self, name=""):
+        from vyper.semantics.types.user.struct import StructDefinition
+
+        if isinstance(self.value_type, StructDefinition):
+            # For structs, set `components` as the struct's components directly
+            return {
+                "name": name,
+                "type": self.json_abi_type,
+                "components": self.value_type.json_abi()["components"],
+            }
+
+        elif isinstance(self.value_type, (ArrayDefinition, DynamicArrayDefinition)):
+            # For static and dynamic arrays, set `components` if the value type is a struct.
+            # Otherwise, the `canonical_abi_type` is sufficient.
+            ret = {
+                "name": name,
+                "type": self.json_abi_type,
+            }
+
+            struct_components = _generate_components_if_struct(self.value_type)
+            if struct_components is not None:
+                ret["components"] = struct_components
+
+            return ret
+
+        return {"name": name, "type": self.json_abi_type}
 
     @property
     def is_dynamic_size(self):
@@ -276,6 +351,12 @@ class TupleDefinition(_SequenceDefinition):
     @property
     def abi_type(self) -> ABIType:
         return ABI_Tuple([t.abi_type for t in self._member_types])
+
+    def json_abi(self, name=""):
+        return {
+            "type": "tuple",
+            "components": [i.json_abi() for i in self.value_type],
+        }
 
     @property
     def size_in_bytes(self):
